refactor(accessible-controller): more testable & idiomatic
- move code into modules
- introduce `VirtualKeyboard` trait to allow mocking device in tests
- use arrays instead of `Vec` for fixed-length buffers
